requirements.txt
# requirements.txt

Flask==2.2.3
tensorflow
numpy
Pillow
gunicorn
Flask-Cors